feat: Update Main.java and WebServerConfig.java
This commit updates the code in Main.java and WebServerConfig.java. In Main.java, the file path for the WebServerConfig object is modified to point to the correct directory. In WebServerConfig.java, the openBrowser() method is updated to include the "/index.html" path in the URL. Additionally, the MyHandler.java file is modified to set the content type as "text/javascript" for JavaScript files. Finally, the index.html file is updated to include a favicon link.